requirements.txt
praw==7.7.1
sqlalchemy==2.0.41
pandas==2.2.2
python-dotenv==1.1.0